+++ lib/domain/repository/settings.dart
@@ -87,4 +87,7 @@ abstract class AbstractSettingsRepository {
 
   /// Sets the [ApplicationSettings.pinnedActions] value.
   Future<void> setPinnedActions(List<String> buttons);
+
+  /// Sets the [ApplicationSettings.callButtonsPosition] value.
+  Future<void> setCallButtonsPosition(CallButtonsPosition position);
 }