+++ lib/store/chat.dart
@@ -40,6 +40,7 @@ import '/domain/model/chat_item_quote_input.dart' as model;
 import '/domain/model/chat_message_input.dart' as model;
 import '/domain/model/mute_duration.dart';
 import '/domain/model/native_file.dart';
+import '/domain/model/ongoing_call.dart';
 import '/domain/model/precise_date_time/precise_date_time.dart';
 import '/domain/model/sending_status.dart';
 import '/domain/model/user.dart';
@@ -211,6 +212,11 @@ class ChatRepository extends DisposableInterface
   /// Indicates whether this [ChatRepository] uses a remote pagination.
   @visibleForTesting
   bool get isRemote => _localPagination == null && _pagination != null;
+
+  /// Returns the map of the current [OngoingCall]s.
+  ///
+  /// Used for [RxChat.inCall] indicator.
+  RxObsMap<ChatId, Rx<OngoingCall>> get calls => _callRepo.calls;
 
   @override
   Future<void> init({